Add support for decoding SIRC (Sony) protocol.
This requires a slightly different implementation of the pulse decoding to
make sense of the commands.

The mark/space coding might actually be the same between the two and I'm
not sure what's the best choice to implement this, between the binning and
the guessing.

Note that with this code in, adafruit_irremote.py needs to be compiled to
mpy for the Adafruit Feather M0 to be able to load.